Adds naive deletion confirmation

+++ app/models/account.js
@@ -14,8 +14,12 @@ export default DS.Model.extend({
     return this.save(); // return so we can work off of the promise in the route
   },
 
-  remove() { // this is where I want to deal with the 'account'
+  remove(model) { // this is where I want to deal with the 'account'
     console.log(`remove() in 'account' model`);
-    return this.destroyRecord(); // talk about delete record here...
+    const userConfirmed = confirm(`Are you sure you want to delete the ${model.data.name} record?`);
+    if (userConfirmed) {
+      return this.destroyRecord(); // talk about delete record here...
+    }
+    return; // $question - .then is not a function if not confirmed / explain
   },
 });
